[bugfix] reorganize the module such that delayed evaluation of type does not interfere with the typing hint value (class vs string) returned by the inspect
