+++ sorter_node_editor.py
@@ -1,17 +1,17 @@
 from ctypes import alignment
+
+import dearpygui.dearpygui as dpg
+
+import pyneurode as dc
+from pyneurode.node_editor import node_editor
+from pyneurode.processor_node.AnalogVisualizer import *
 from pyneurode.processor_node.GUIProcessor import GUIProcessor
+from pyneurode.processor_node.MountainsortTemplateProcessor import \
+    MountainsortTemplateProcessor
 from pyneurode.processor_node.Processor import *
 from pyneurode.processor_node.ProcessorContext import ProcessorContext
-from pyneurode.processor_node.FileReaderSource import FileReaderSource
-from pyneurode.processor_node.SpikeSortProcessor import SpikeSortProcessor
-from pyneurode.processor_node.SyncDataProcessor import SyncDataProcessor
-import pyneurode as dc
-import logging
-import dearpygui.dearpygui as dpg
-from pyneurode.processor_node.AnalogVisualizer import *
-import inspect
-import dearpygui.dearpygui as dpg
-from pyneurode.node_editor import node_editor
+from pyneurode.processor_node.ScaleProcessor import ScaleProcessor
+from pyneurode.processor_node.SineTimeSource import SineTimeSource
 
 dpg.create_context()
 
@@ -25,45 +25,45 @@ have one single GUI node in the main thread where all the others can connect to
 
 '''
 
-def make_node(processor:Processor) -> Tuple[Dict, Dict]:
-    '''
-    Parse the object information and make it into a node
-    '''
-    with dpg.node(label=processor.proc_name):
-        sig = inspect.signature(processor.__init__)
+# def make_node(processor:Processor) -> Tuple[Dict, Dict]:
+#     '''
+#     Parse the object information and make it into a node
+#     '''
+#     with dpg.node(label=processor.proc_name):
+#         sig = inspect.signature(processor.__init__)
         
-        input_cls, output_cls = processor.get_IOspecs()
+#         input_cls, output_cls = processor.get_IOspecs()
         
         
-        # Add input and output nodes
-        inputs = {}
-        for input in input_cls:
-            with dpg.node_attribute(label=processor.proc_name, attribute_type=dpg.mvNode_Attr_Input) as node_input:
-                dpg.add_text(input.__name__)
-                inputs[input.__name__] = node_input
+#         # Add input and output nodes
+#         inputs = {}
+#         for input in input_cls:
+#             with dpg.node_attribute(label=processor.proc_name, attribute_type=dpg.mvNode_Attr_Input) as node_input:
+#                 dpg.add_text(input.__name__)
+#                 inputs[input.__name__] = node_input
                 
         
-        outputs = {}
-        for output in output_cls:
-            with dpg.node_attribute(label=processor.proc_name, attribute_type=dpg.mvNode_Attr_Output) as node_output:
-                dpg.add_text(output.__name__, indent=200)
-                outputs[output.__name__] = node_output
+#         outputs = {}
+#         for output in output_cls:
+#             with dpg.node_attribute(label=processor.proc_name, attribute_type=dpg.mvNode_Attr_Output) as node_output:
+#                 dpg.add_text(output.__name__, indent=200)
+#                 outputs[output.__name__] = node_output
                 
                 
         
-        for param in sig.parameters.keys():
-            with dpg.node_attribute(label=processor.proc_name, attribute_type=dpg.mvNode_Attr_Static): #static attribute
-                if sig.parameters[param].annotation == 'float':
-                    dpg.add_input_float(label=param, width = 150)
-                elif sig.parameters[param].annotation == 'int':
-                    dpg.add_input_int(label=param, width = 150)
+#         for param in sig.parameters.keys():
+#             with dpg.node_attribute(label=processor.proc_name, attribute_type=dpg.mvNode_Attr_Static): #static attribute
+#                 if sig.parameters[param].annotation == 'float':
+#                     dpg.add_input_float(label=param, width = 150)
+#                 elif sig.parameters[param].annotation == 'int':
+#                     dpg.add_input_int(label=param, width = 150)
                     
-        return inputs, outputs
+#         return inputs, outputs
                     
                     
-def link_callback(sender,app_data):
-    print(f'Connecting {app_data[0]} and {app_data[1]}')
-    dpg.add_node_link(app_data[0], app_data[1], parent=sender)
+# def link_callback(sender,app_data):
+#     print(f'Connecting {app_data[0]} and {app_data[1]}')
+#     dpg.add_node_link(app_data[0], app_data[1], parent=sender)
                     
 
 if  __name__ == '__main__':
@@ -72,9 +72,10 @@ if  __name__ == '__main__':
 
     with ProcessorContext(auto_start=False) as ctx:
 
-        sineWave = SineTimeSource(0.1,frequency = 12.3, channel_num=10, sampling_frequency=100)
+        sineWave = SineTimeSource(interval = 0.1,frequency = 12.3, channel_num=10, sampling_frequency=100)
         scaleProcessor = ScaleProcessor(4,5)
         analog_visualizer = AnalogVisualizer('v1',scale=4,time_scale=1/100)
+
         
         # gui = GUIProcessor()
         # gui.register_visualizer(analog_visualizer,filters=['sine'])
