+++ src/pyneurode/processor_node/Processor.py
@@ -306,50 +306,7 @@ class Message2(Message):
     dtype = 'message2'
     def __init__(self, data: Any, timestamp: Optional[float] = None):
         super().__init__(Message1.dtype, data, timestamp)
-            
-class SineTimeSource(TimeSource):
-    '''
-    A TimeSource that generate multi-channel sine wave for debugging purpose
-    '''
-    def __init__(self, interval:float, frequency:float, channel_num:int, sampling_frequency:int=100,
-                 noise_std:float=0, scale=1, offset=0, msg_type =None):
-        super().__init__(interval)
-        self.frequency = frequency #in Hz
-        self.start_time = time.time()
-        self.channel_num = int(channel_num)
-        self.last_time = None #last time the data are plotted
-        self.dt = 1/sampling_frequency
-        self.noise_std = noise_std
-        self.scale = scale
-        self.offset = offset
-        self.msg_type = msg_type
-
-    def process(self):
-        # Generate a sine waveform
-        now = time.time() - self.start_time
-
-        if self.last_time is None:
-            t = np.arange(0, now, step=self.dt)
-        else:
-            # t = np.linspace(self.last_time, now, int((now-self.last_time)*self.Fs))
-            t = np.arange(self.last_time, now, step=self.dt)
-
-        if len(t)> 0:
-            # print(t)
-            self.last_time = t[-1]+self.dt
-            y = np.sin(2*np.pi*self.frequency*t)*self.scale+self.offset
-            y = np.tile(y, (self.channel_num,1)).T
-            
-            if self.noise_std > 0:
-                # add in noise
-                y += np.random.normal(scale=self.noise_std, size=y.shape)
-                
-            if self.msg_type is None:
-                return Message('sine', y)
-            elif self.msg_type =='message1':
-                return Message1(y)
-            else:
-                return Message2(y)
+ 
                                 
 
 class DummyTimeSource(TimeSource):
@@ -384,16 +341,7 @@ class MsgTimeSource(TimeSource):
 class AddProcessor(Processor):
     def process(self, msg):
         return Message('dummy',msg.data + 100)
-    
-class ScaleProcessor(Processor):
-    def __init__(self, a:float, b:float):
-        super().__init__()
-        self.a = a
-        self.b = b
-        
-    def process(self, message: Message) -> Message:
-        x = message.data*self.a+self.b
-        return Message(message.type, x)
+
 
 class EchoSink(Sink):
     """ Simply print the message it receives
